+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/CRUD/Read/FilterByMasterFieldTest.cs
@@ -17,12 +17,12 @@
     /// </summary>
 #if NETFRAMEWORK
     public class FilterByMasterFieldTest : BaseODataServiceIntegratedTest
-#endif
-#if NETCOREAPP
+#else
     public class FilterByMasterFieldTest : BaseODataServiceIntegratedTest<TestStartup>
 #endif
     {
-#if NETCOREAPP
+#if NETFRAMEWORK
+#else
         /// <summary>
         /// Конструктор по-умолчанию.
         /// </summary>