+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/CRUD/Update/WebFileTest.cs
@@ -15,12 +15,12 @@
 
 #if NETFRAMEWORK
     public class WebFileTest : BaseODataServiceIntegratedTest
-#endif
-#if NETCOREAPP
+#else
     public class WebFileTest : BaseODataServiceIntegratedTest<TestStartup>
 #endif
     {
-#if NETCOREAPP
+#if NETFRAMEWORK
+#else
         /// <summary>
         /// Конструктор по-умолчанию.
         /// </summary>