+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/Events/AfterGetTest.cs
@@ -13,12 +13,12 @@
 
 #if NETFRAMEWORK
     public class AfterGetTest : BaseODataServiceIntegratedTest
-#endif
-#if NETCOREAPP
+#else
     public class AfterGetTest : BaseODataServiceIntegratedTest<TestStartup>
 #endif
     {
-#if NETCOREAPP
+#if NETFRAMEWORK
+#else
         /// <summary>
         /// Конструктор по-умолчанию.
         /// </summary>