+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/Events/AfterSaveTest.cs
@@ -25,12 +25,12 @@
     /// </summary>
 #if NETFRAMEWORK
     public class AfterSaveTest : BaseODataServiceIntegratedTest
-#endif
-#if NETCOREAPP
+#else
     public class AfterSaveTest : BaseODataServiceIntegratedTest<TestStartup>
 #endif
     {
-#if NETCOREAPP
+#if NETFRAMEWORK
+#else
         /// <summary>
         /// Конструктор по-умолчанию.
         /// </summary>