+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/XUnitTestRunnerInitializer.cs
@@ -2,7 +2,8 @@
 
 namespace NewPlatform.Flexberry.ORM.IntegratedTests
 {
-#if NETCOREAPP
+#if NETFRAMEWORK
+#else
     using System.Configuration;
     using System.IO;
     using System.Reflection;
@@ -30,7 +31,8 @@ namespace NewPlatform.Flexberry.ORM.IntegratedTests
         public XUnitTestRunnerInitializer(IMessageSink messageSink)
             : base(messageSink)
         {
-#if NETCOREAPP
+#if NETFRAMEWORK
+#else
             Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
             string configFile = $"{Assembly.GetExecutingAssembly().Location}.config";
             string outputConfigFile = ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.None).FilePath;